% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oXim-main.R
\docType{data}
\name{colPalette}
\alias{colPalette}
\title{Default color palette most using on acostic echograms.}
\format{A vector of 256 colors in RBG format.}
\usage{
colPalette
}
\description{
Vector with 256 colors commonly used on echograms plots.
}
\references{
Boletines del Instituto del Mar del Peru.
}

