% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oXim-main.R
\docType{data}
\name{defaultFilterSettings}
\alias{defaultFilterSettings}
\title{Default set of filter parameters.}
\format{A \code{data.frame} with columns name, type, radius, times and tolerance.}
\usage{
defaultFilterSettings
}
\description{
\code{data.frame} object containig a settings used by \code{getOxyrange}
function to apply echograms' matrix. It can be usefull as example of filter settings.
}
