% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoData-class.R
\name{echogramPlot.echoData}
\alias{echogramPlot.echoData}
\title{echogramPlot method for echoData}
\usage{
\method{echogramPlot}{echoData}(x, ...)
}
\arguments{
\item{x}{Object of class \code{echoData}}

\item{...}{Extra arguments passed to \code{\link{echogramPlot}} function.}
}
\description{
This method takes an \code{echoData} object and plots output echograms.
}
