% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxyclineData-class.R
\name{echogramPlot.oxyclineData}
\alias{echogramPlot.oxyclineData}
\title{echogramPlot method for oxyclineData}
\usage{
\method{echogramPlot}{oxyclineData}(x, colEchogram = "colPalette",
  oxyLine = TRUE, oxyLineParams = list(), ...)
}
\arguments{
\item{x}{Object of class \code{oxyclineData} with internal echogram matrix to be plotted.}

\item{colEchogram}{Pallete of colours to plot the echograms. If \code{NULL} (default) the system
will use the same combination used on object \code{colPallete}.}

\item{oxyLine}{\code{logical}. Do you want to add oxycline line to the plot?}

\item{oxyLineParams}{If \code{oxyLine = TRUE}, parameters passed to \code{\link{lines}} function.}

\item{...}{Extra arguments passed to \code{\link{echogramPlot}} function.}
}
\description{
This method takes an \code{oxyclineData} object and plots output echograms. Optionaly,
users can add oxycline line.
}

