% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oXim-main.R
\docType{package}
\name{oXim-package}
\alias{oXim}
\alias{oXim-package}
\title{Oxycline Index from Matrix Echograms}
\description{
Tools for oxycline depth calculation from echogram matrices
}
\author{
Wencheng Lau-Medrano, \email{luis.laum@gmail.com}
}
\keyword{depth,}
\keyword{echograms,}
\keyword{image-filtering}
\keyword{oxycline,}

