% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoData-class.R
\name{plot.echoData}
\alias{plot.echoData}
\title{Plot method for echoData Objects.}
\usage{
\method{plot}{echoData}(x, ...)
}
\arguments{
\item{x}{\code{echoData} object provided by \code{readEchograms} function.}

\item{...}{Arguments passed to \code{\link{echogramPlot}} function.}
}
\description{
This function takes an \code{echoData} object and plots an interpolated map
showing oxycline depth along shore.
}

