% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoData-class.R
\name{print.echoData}
\alias{print.echoData}
\title{Print method for echoData Objects.}
\usage{
\method{print}{echoData}(x, ...)
}
\arguments{
\item{x}{\code{echoData} object provided by \code{readEchograms} function.}

\item{...}{Extra argumemts.}
}
\description{
Shows main information from echodata Objects.
}
