% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxyclineData-class.R
\name{print.oxyclineData}
\alias{print.oxyclineData}
\title{Print method for oxyclineData Objects.}
\usage{
\method{print}{oxyclineData}(x, ...)
}
\arguments{
\item{x}{Object of class \code{oxyclineData}.}

\item{...}{Extra argumemts.}
}
\description{
Shows main information from oxyclineData Objects.
}

