% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoData-class.R
\name{print.summary.echoData}
\alias{print.summary.echoData}
\title{Print method for summary.echoData}
\usage{
\method{print}{summary.echoData}(x, ...)
}
\arguments{
\item{x}{\code{echoData.summary} object provided by application of summary method to \code{echoData} object.}

\item{...}{Extra argumemts.}
}
\description{
Shows main information from echodata.summary Objects.
}

