% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxyclineData-class.R
\name{print.summary.oxyclineData}
\alias{print.summary.oxyclineData}
\title{Print method for summary.oxyclineData}
\usage{
\method{print}{summary.oxyclineData}(x, ...)
}
\arguments{
\item{x}{Object of class \code{summary.oxyclineData}.}

\item{...}{Extra argumemts.}
}
\description{
Shows main information from \code{oxyclineData.summary} objects.
}
