% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoData-class.R
\name{summary.echoData}
\alias{summary.echoData}
\title{Summary method for echoData}
\usage{
\method{summary}{echoData}(object, ...)
}
\arguments{
\item{object}{\code{echoData} object provided by \code{readEchograms} function.}

\item{...}{Extra argumemts.}
}
\description{
Get summary information of echograms included on echodata Objects.
}
