% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxyclineData-class.R
\name{summary.oxyclineData}
\alias{summary.oxyclineData}
\title{Summary method for oxyclineData Objects.}
\usage{
\method{summary}{oxyclineData}(object, ...)
}
\arguments{
\item{object}{Object of class \code{oxyclineData}.}

\item{...}{Extra argumemts.}
}
\description{
Get summary information of oxycline depth limits included on \code{oxyclineData} objects.
}

