% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oXim-main.R
\docType{package}
\name{oXim-package}
\alias{oXim}
\alias{oXim-package}
\title{Oxycline Index from Matrix Echograms}
\description{
Pack oriented to extract oxycline depth from echogram matrix using median-filter and 2D-convolution based algorithms.
}
\author{
Wencheng Lau-Medrano, \email{luis.laum@gmail.com}
}
\references{
oXim: Oxycline Index from Matrix Echograms (RJournal)
}
\keyword{depth,}
\keyword{echograms,}
\keyword{image-filtering}
\keyword{oxycline,}

